Bugfixes according to review

+++ models/contacts.js
@@ -11,7 +11,9 @@ const listContacts = async userId => {
 
 const getContactById = async (userId, contactId) => {
   try {
-    return await Contact.findOne({ owner: userId, _id: contactId });
+    const contact = await Contact.findOne({ owner: userId, _id: contactId });
+    if (!contact) return 404;
+    return contact;
   } catch (err) {
     console.log(`Error getting contact with id ${contactId}: `, err);
     throw err;
@@ -20,6 +22,8 @@ const getContactById = async (userId, contactId) => {
 
 const removeContact = async (contactId, userId) => {
   try {
+    const contact = await Contact.findOne({ owner: userId, _id: contactId });
+    if (!contact) return 404;
     return await Contact.findByIdAndRemove({ owner: userId, _id: contactId });
   } catch (err) {
     console.log(`Error removing contact with id ${contactId}: `, err);
@@ -41,7 +45,10 @@ const addContact = async (body, userId) => {
 };
 
 const updateContact = async (contactId, body, userId) => {
+  console.log('Update Contact UserID:', userId);
   try {
+    const contact = await Contact.findOne({ owner: userId, _id: contactId });
+    if (!contact) return 404;
     return await Contact.findByIdAndUpdate({ owner: userId, _id: contactId }, body, { new: true });
   } catch (err) {
     console.error('An error occurred while updating contact: ', err);
@@ -51,6 +58,8 @@ const updateContact = async (contactId, body, userId) => {
 
 const updatedStatusContact = async (contactId, favorite, userId) => {
   try {
+    const contact = await Contact.findOne({ owner: userId, _id: contactId });
+    if (!contact) return 404;
     return await Contact.findByIdAndUpdate(
       { owner: userId, _id: contactId },
       { $set: { favorite: favorite } },
